move calculate_foldchange to StatisticUtils
